credential-management/fedcm-button-mode-basics.https.html

<!DOCTYPE html>
<title>Federated Credential Management API auto selected flag tests.</title>
<link rel="help" href="https://fedidcg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<script type="module">
import {request_options_with_mediation_required,
        fedcm_test,
        select_manifest,
        fedcm_get_and_select_first_account} from './support/fedcm-helper.sub.js';

fedcm_test(async t => {
  let test_options = request_options_with_mediation_required();
  test_options.identity.mode = "button";

  let result = navigator.credentials.get(test_options);
  return promise_rejects_dom(t, 'NetworkError', result);
}, "Test that the button mode without user activation will fail.");

fedcm_test(async t => {
  let test_options = request_options_with_mediation_required();
  test_options.identity.mode = "button";

  await test_driver.bless('user activation');

  let cred = await fedcm_get_and_select_first_account(t, test_options);
  assert_equals(cred.token, "token");
}, "Test that the button mode consumes user activation.");

fedcm_test(async t => {
  let test_options = request_options_with_mediation_required("manifest_with_rp_mode.json");
  await select_manifest(t, test_options);
  test_options.identity.mode = "button";

  await test_driver.bless('user activation');

  let cred = await fedcm_get_and_select_first_account(t, test_options);
  assert_equals(cred.token, "mode=button");
}, "Test that the rp mode bit is properly sent in button flow.");

</script>
